% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funky.gen.R
\name{config}
\alias{config}
\title{Package configuration data}
\usage{
config(pkg = utils::packageName(env = parent.frame()))
}
\arguments{
\item{pkg}{\R package name. A character scalar. Defaults to the name of the calling package.}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with the columns \code{key}, \code{default_value}, \code{default_value_dynamic}, \code{require}, and \code{description}.
}
\description{
Retrieves a package's funky configuration data (\verb{<pkg>:::funky_config}).
}
\examples{
try(
  funky::config(pkg = "pkgpurl")
)
}
\seealso{
Other package configuration data functions:
\code{\link{augment_config}()},
\code{\link{print_config}()}
}
\concept{pkg_config_data}
