% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funky.gen.R
\name{config_val_default}
\alias{config_val_default}
\title{Get default package configuration value}
\usage{
config_val_default(
  key,
  pkg = utils::packageName(env = parent.frame()),
  env = parent.frame()
)
}
\arguments{
\item{key}{Configuration key name. A character scalar.}

\item{pkg}{\R package name. A character scalar. Defaults to the name of the calling package.}

\item{env}{Environment to evaluate \code{default_value_dynamic} in, if necessary.}
}
\value{
\code{pkgsnip::return_lbl("r_obj")}
}
\description{
Retrieves a package configuration key's default value from the package's funky configuration data (column \code{default_value} or \code{default_value_dynamic} of
\verb{<pkg>:::funky_config}). If no default value is specified (\code{NULL}), nothing is returned (\code{NULL}).
}
\examples{
try(
  funky::config_val_default(key = "gen_pkgdown_ref",
                            pkg = "pkgpurl")
)
}
\seealso{
Other package configuration value functions:
\code{\link{config_val}()},
\code{\link{has_config_val}()}
}
\concept{pkg_config_vals}
