% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funky.gen.R
\name{env_var_name}
\alias{env_var_name}
\title{Assemble environment variable name}
\usage{
env_var_name(key, pkg = utils::packageName(env = parent.frame()))
}
\arguments{
\item{key}{Configuration key name. A character scalar.}

\item{pkg}{\R package name. A character scalar. Defaults to the name of the calling package.}
}
\value{
A character scalar.
}
\description{
Assembles the environment variable name for a given package configuration key.
}
\examples{
funky:::env_var_name(key = "gen_pkgdown_ref",
                     pkg = "pkgpurl")
}
\keyword{internal}
