% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funky-package.R
\docType{package}
\name{funky-package}
\alias{funky}
\alias{funky-package}
\title{funky: Canonicalized R package configuration}
\description{
A canonicalized way to R package configuration striving for user and developer convenience alike.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://funky.rpkg.dev}
  \item \url{https://gitlab.com/rpkg.dev/funky}
  \item Report bugs at \url{https://gitlab.com/rpkg.dev/funky/-/issues}
}

}
\author{
\strong{Maintainer}: Salim Brüggemann \email{salim@rpkg.dev} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID})

}
\keyword{internal}
