% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funky.R
\name{get_call_args_helper}
\alias{get_call_args_helper}
\title{Helper function for \code{\link[=get_call_args]{get_call_args()}}}
\usage{
get_call_args_helper(n = 1L, err_h)
}
\arguments{
\item{n}{The number ob frames to go back. The default value \code{n = 0L} means,
that the given arguments of current function should be returned.}

\item{err_h}{An error handling function created with \code{\link[composerr:composerr]{composerr::composerr()}}.}
}
\description{
Helper function for \code{\link[=get_call_args]{get_call_args()}}
}
