% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funky.gen.R
\name{opt_name}
\alias{opt_name}
\title{Assemble \R option name}
\usage{
opt_name(key, pkg)
}
\arguments{
\item{key}{Configuration key name. A character scalar.}

\item{pkg}{\R package name. A character scalar. Defaults to the name of the calling package.}
}
\value{
A character scalar.
}
\description{
Assembles the \R option name for a given package configuration key.
}
\examples{
funky:::opt_name(key = "gen_pkgdown_ref",
                 pkg = "pkgpurl")
}
\keyword{internal}
