% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funky.gen.R
\name{print_config}
\alias{print_config}
\title{Print package configuration data}
\usage{
print_config(
  pkg = utils::packageName(env = parent.frame()),
  roxy_to_md = FALSE
)
}
\arguments{
\item{pkg}{\R package name. A character scalar. Defaults to the name of the calling package.}

\item{roxy_to_md}{Whether or not to convert roxygen2 documentation \href{https://roxygen2.r-lib.org/articles/rd-formatting.html#function-links}{links in pseudo-Markdown style} to actual Markdown ones using \code{\link[pal:roxy_to_md_links]{pal::roxy_to_md_links()}}.}
}
\value{
A character vector.
}
\description{
Prints a package's funky configuration data (\verb{<pkg>:::funky_config}) as a prettily formatted Markdown table.
}
\examples{
try(
  funky::print_config(pkg = "pkgpurl")
)
}
\seealso{
Other package configuration data functions:
\code{\link{augment_config}()},
\code{\link{config}()}
}
\concept{pkg_config_data}
