% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funky.gen.R
\docType{data}
\name{ptype_config}
\alias{ptype_config}
\title{Package configuration data prototype}
\format{
A \link[tibble:tbl_df-class]{tibble} with the columns \code{key}, \code{default_value}, \code{default_value_dynamic}, \code{require}, and \code{description}.
}
\usage{
ptype_config
}
\description{
An empty funky configuration data \link[tibble:tbl_df-class]{tibble}. E.g. useful to start from when defining funky configuration data via \code{\link[tibble:add_row]{tibble::add_row()}} calls.
}
\examples{
funky::ptype_config
}
\seealso{
Other package configuration data functions:
\code{\link{augment_config}()},
\code{\link{config}()},
\code{\link{print_config}()}
}
\concept{pkg_config_data}
\keyword{datasets}
